% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/casewise.R
\name{back2timereg}
\alias{back2timereg}
\title{Convert to timereg object}
\usage{
back2timereg(obj)
}
\arguments{
\item{obj}{no use}
}
\description{
convert to timereg object
}
\author{
Thomas Scheike
}

