% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{base1cumhaz}
\alias{base1cumhaz}
\title{rate of CRBSI for HPN patients of Copenhagen}
\source{
Estimated data
}
\description{
rate of CRBSI for HPN patients of Copenhagen
}
\keyword{data}
