% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{base2cumhaz}
\alias{base2cumhaz}
\title{rate of Mechanical complication for catheter of HPN patients of Copenhagen}
\source{
Estimated data
}
\description{
rate of Mechanical complication for catheter of HPN patients of Copenhagen
}
\keyword{data}
