% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{base44cumhaz}
\alias{base44cumhaz}
\title{rate of Occlusion/Thrombosis complication for catheter of HPN patients of Copenhagen}
\source{
Estimated data
}
\description{
rate of Occlusion/Thrombosis complication for catheter of HPN patients of Copenhagen
}
\keyword{data}
