\name{bptwin-package}
\alias{bptwin-package}
\docType{package}
\title{
  Bivariate probit models for twin data
}
\description{
  Bivariate probit model/Liability model with right censoring and left truncation
}
\details{
\tabular{ll}{
Package: \tab bptwin\cr
Type: \tab Package\cr
Version: \tab 0.6-3\cr
Date: \tab 2011-02-07\cr
License: \tab GPL (>=2)\cr
URL: http://r-forge.r-project.org/projects/lava/
LazyLoad: \tab yes\cr
}
bptwin
ipw
cumh
cif
prepcif
fastapprox
bpnd
}
\author{
Klaus K. Holst <k.k.holst@biostat.ku.dk>
}
\keyword{ package }
\examples{
example(bptwin)
}
