\name{ClaytonOakes}
\alias{ClaytonOakes}
\title{Clayton-Oakes model with piece-wise constant hazards}
\usage{
  ClaytonOakes(formula, data = parent.frame(), id,
    var.formula = ~1, cuts = NULL, type = "co", start,
    control = list(), ...)
}
\arguments{
  \item{formula}{}

  \item{data}{}

  \item{id}{}

  \item{var.formula}{}

  \item{cuts}{}

  \item{type}{}

  \item{start}{}

  \item{control}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

