% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{cluster.index}
\alias{cluster.index}
\title{Finds subjects related to same cluster}
\usage{
cluster.index(clusters, index.type = FALSE, num = NULL, Rindex = 0,
  mat = NULL, return.all = FALSE, code.na = NA)
}
\description{
Finds subjects related to same cluster
}
\examples{
i<-c(1,1,2,2,1,3)
d<- cluster.index(i)
print(d)
}
\author{
Klaus Holst, Thomas Scheike
}
\references{
Cluster indeces
}
\keyword{cluster}
\keyword{indeces}

