\name{coef.biprobit}
\alias{coef.biprobit}
\title{Extract parameter estimates from biprobit object}
\usage{
  coef.biprobit(object, matrix = FALSE, ...)
}
\arguments{
  \item{object}{}

  \item{matrix}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

