\name{coef.biprobit.strata}
\alias{coef.biprobit.strata}
\title{Extract parameter estimates of biprobit.strata objects}
\usage{
  coef.biprobit.strata(object, ...)
}
\arguments{
  \item{object}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

