\name{conc2probandwise}
\alias{conc2probandwise}
\title{Estimates the probandwise concordance based on Concordance and marginal estimate}
\usage{
  conc2probandwise(conc, marg, test = "no-test")
}
\arguments{
  \item{conc}{Concordance}

  \item{marg}{Marginal estimate}

  \item{test}{Type of test for independence assumption.
  "conc" makes test on concordance scale and "case" means a
  test on the probandwise concordance}
}
\description{
  .. content for description (no empty lines) ..
}
\author{
  Thomas Scheike
}

