\name{concordance}
\alias{concordance}
\title{Concordance}
\usage{
  concordance(object, cif1, cif2 = NULL, messages = TRUE,
    model = NULL, coefs = NULL, ...)
}
\arguments{
  \item{object}{}

  \item{cif1}{}

  \item{cif2}{}

  \item{messages}{}

  \item{model}{}

  \item{coefs}{}

  \item{...}{}
}
\description{
  Concordance
}
\author{
  Thomas Scheike
}

