% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aalen.r
\name{const}
\alias{const}
\title{Identifies parametric terms of model}
\usage{
const(x)
}
\arguments{
\item{x}{variable}
}
\description{
Specifies which of the regressors that have constant effect.
}
\author{
Thomas Scheike
}
\keyword{survival}
