% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.prop-excess.r
\name{cox}
\alias{cox}
\title{Identifies proportional excess terms of model}
\usage{
cox(x)
}
\arguments{
\item{x}{variable}
}
\description{
Specifies which of the regressors that lead to proportional excess hazard
}
\author{
Thomas Scheike
}
\keyword{survival}
