\name{cumh}
\alias{cumh}
\title{Calculate}
\usage{
  cumh(formula, data, ..., time,
    timestrata = quantile(data[, time], c(0.25, 0.5, 0.75, 1)),
    cumulative = TRUE, silent = FALSE)
}
\arguments{
  \item{formula}{}

  \item{data}{}

  \item{...}{}

  \item{time}{}

  \item{timestrata}{}

  \item{cumulative}{}

  \item{silent}{}
}
\description{
  .. content for description (no empty lines) ..
}
\details{
  .. content for details ..
}
\author{
  Klaus K. Holst and Thomas Scheike
}

