% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{daggregate}
\alias{daggregate}
\title{aggregating for for data frames}
\usage{
daggregate(data, x, ..., group = NULL, fun = "summary", regex = FALSE)
}
\description{
aggregating for for data frames
}
\examples{
data("sTRACE",package="timereg")
daggregate(iris, "^.e.al", group="Species", fun=cor, regex=TRUE)
daggregate(iris, ~Sepal.Length+Petal.Length|Species, fun=summary)
daggregate(iris, ~log(Sepal.Length)+I(Petal.Length>1.5)|Species, fun=summary)
daggregate(iris, "*Length*", group="Species", fun=head)
daggregate(iris, "^.e.al", group="Species", fun=tail, regex=TRUE)
daggregate(sTRACE, status~ diabetes, fun=table)
daggregate(sTRACE, status~ diabetes+sex, fun=table)
daggregate(sTRACE, status~ diabetes+sex|vf+I(wmi>1.4), fun=table)
daggregate(iris, "^.e.al", group="Species")
daggregate(iris, I(Sepal.Length>7)~Species | I(Petal.Length>1.5))
daggregate(iris, I(Sepal.Length>7)~Species | I(Petal.Length>1.5), fun=table)
daggregate(iris, I(Sepal.Length>7)~Species | I(Petal.Length>1.5), fun=lava:::Print)
}

