% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Dbvn}
\alias{Dbvn}
\title{Derivatives of the bivariate normal cumulative distribution function}
\usage{
Dbvn(p,design=function(p,...) {
     return(list(mu=cbind(p[1],p[1]),
               dmu=cbind(1,1),
               S=matrix(c(p[2],p[3],p[3],p[4]),ncol=2),
               dS=rbind(c(1,0,0,0),c(0,1,1,0),c(0,0,0,1)))  )},
     Y=cbind(0,0))
}
\arguments{
\item{p}{Parameter vector}

\item{design}{Design function with defines mean, derivative of mean, variance,
and derivative of variance with respect to the parameter p}

\item{Y}{column vector where the CDF is evaluated}
}
\description{
Derivatives of the bivariate normal cumulative distribution function
}
\author{
Klaus K. Holst
}

