\docType{data}
\name{dermalridgesMZ}
\alias{dermalridgesMZ}
\title{Dermal ridges data (monozygotic twins)}
\source{
  Sarah B. Holt (1952). Genetics of dermal ridges:
  bilateral asymmetry in finger ridge-counts.  Annals of
  Eugenics 17 (1), pp 211--231. DOI:
  10.1111/j.1469-1809.1952.tb02513.x
}
\description{
  Data on dermal ridge counts in left and right hand in
  (nuclear) families
}
\keyword{data}
\keyword{datasets}

