% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{drcumhaz}
\alias{drcumhaz}
\title{Rate for leaving HPN program for patients of Copenhagen}
\source{
Estimated data
}
\description{
Rate for leaving HPN program for patients of Copenhagen
}
\keyword{data}
