% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{dtable}
\alias{dtable}
\title{tables for data frames}
\usage{
dtable(data, x, g, all2by2 = TRUE, regex = FALSE, ...)
}
\arguments{
\item{data}{if x is formula or names for data frame then data frame is needed.}

\item{x}{name of variable, or fomula, or names of variables on data frame.}

\item{g}{possible group variable}

\item{...}{Optional additional arguments}
}
\description{
tables for data frames
}
\examples{
data("sTRACE",package="timereg")
dt<- sTRACE

dtable(dt,~status+vf,~chf+diabetes)

dtable(dt,status+vf~chf+diabetes)

dtable(dt,c("*f*","status"),~diabetes)
dtable(dt,c("*f*","status"),~diabetes,all2by2=FALSE)

}
\author{
Klaus K. Holst and Thomas Scheike
}

