% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtransform.R
\name{dtransform}
\alias{dtransform}
\alias{dtransform<-}
\alias{dtrans}
\alias{dtrans<-}
\title{Transform that allows condition}
\usage{
dtransform(data, ...)
}
\arguments{
\item{data}{is data frame}

\item{...}{new variable definitions including possible if condition}
}
\description{
Defines new variables under condition for data frame
}
\examples{
data(mena)

xx <- dtransform(mena,ll=log(agemena)+twinnum)

xx <- dtransform(mena,ll=log(agemena)+twinnum,agemena<15)
xx <- dtransform(xx  ,ll=100+agemena,ll2=1000,agemena>15)
dsummary(xx,ll+ll2~I(agemena>15))
}
