% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.r
\name{Event}
\alias{Event}
\alias{as.character.Event}
\alias{as.matrix.Event}
\alias{[.Event}
\alias{format.Event}
\alias{print.Event}
\alias{rbind.Event}
\alias{summary.Event}
\title{Event history object}
\usage{
Event(time, time2 = TRUE, cause = NULL, cens.code = 0, ...)
}
\arguments{
\item{time}{Time}

\item{time2}{Time 2}

\item{cause}{Cause}

\item{cens.code}{Censoring code (default 0)}

\item{...}{Additional arguments}
}
\value{
Object of class Event (a matrix)
}
\description{
Constructur for Event History objects
}
\details{
... content for details
}
\examples{

	t1 <- 1:10
	t2 <- t1+runif(10)
	ca <- rbinom(10,2,0.4)
	(x <- Event(t1,t2,ca))

}
\author{
Klaus K. Holst and Thomas Scheike
}
