% Generated by roxygen2 (4.0.1): do not edit by hand
\name{familycluster.index}
\alias{familycluster.index}
\title{Finds all pairs within a cluster (family)}
\usage{
familycluster.index(clusters, index.type = FALSE, num = NULL, Rindex = 1)
}
\arguments{
\item{clusters}{list of indeces}

\item{index.type}{argument of cluster index}

\item{num}{num}

\item{Rindex}{index starts with 1 in R, and 0 in C}
}
\description{
Finds all pairs within a cluster (family)
}
\examples{
i<-c(1,1,2,2,1,3)
d<- familycluster.index(i)
print(d)
}
\author{
Klaus Holst, Thomas Scheike
}
\references{
Cluster indeces
}
\seealso{
cluster.index familyclusterWithProbands.index
}

