% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{familyclusterWithProbands.index}
\alias{familycluster.index}
\alias{familyclusterWithProbands.index}
\title{Finds all pairs within a cluster (famly)  with the proband (case/control)}
\usage{
familyclusterWithProbands.index(clusters, probands, index.type = FALSE,
  num = NULL, Rindex = 1)
}
\description{
second column of pairs are the probands and the first column the related subjects
}
\examples{
i<-c(1,1,2,2,1,3)
p<-c(1,0,0,1,0,1)
d<- familyclusterWithProbands.index(i,p)
print(d)
}
\author{
Klaus Holst, Thomas Scheike
}
\references{
Cluster indeces
}
\keyword{cluster}
\keyword{indeces}

