\name{fast.pattern}
\alias{fast.pattern}
\title{Fast pattern}
\usage{
fast.pattern(x, y, ...)
}
\arguments{
  \item{x}{Matrix (binary) of patterns. Optionally if
  \code{y} is also passed as argument, then the pattern
  matrix is defined as the elements agreeing in the two
  matrices.}

  \item{y}{Optional matrix argument with same dimensions as
  \code{x} (see above)}

  \item{...}{Optional additional arguments}
}
\description{
Fast pattern
}
\author{
Klaus K. Holst
}

