\name{fastapprox}
\alias{fastapprox}
\title{Fast approximation}
\usage{
  fastapprox(t1, t2, y1)
}
\arguments{
  \item{t1}{}

  \item{t2}{}

  \item{y1}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

