% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof-phreg.R
\name{gofG.phreg}
\alias{gofG.phreg}
\title{Stratified baseline graphical GOF test for Cox covariates in PH regression}
\usage{
gofG.phreg(x, coefcox, sim = 1, silent = 1, ...)
}
\arguments{
\item{silent}{to keep it absolutely silent}

\item{...}{Additional arguments to lower level funtions}

\item{formula}{formula for cox regression}

\item{data}{data for model}

\item{offset}{offset}

\item{weights}{weights}

\item{modelmatrix}{matrix for cumulating residuals}

\item{n.sim}{number of simulations for bands on differences}
}
\description{
Looks at stratified baseline in Cox model and plots all baselines versus each
other to see if lines are straight, with 50 resample versions under the 
assumptiosn that the stratified Cox is correct
}
\examples{
data(TRACE)

m1 <- phreg1(Surv(time,status==9)~strata(vf)+chf+wmi,data=TRACE) 
m2 <- phreg1(Surv(time,status==9)~vf+strata(chf)+wmi,data=TRACE) 
par(mfrow=c(2,2))
gofG.phreg(m1)
gofG.phreg(m2)

}
\author{
THomas Scheike and Klaus K. Holst
}
