\name{pval}
\alias{summary.restricted.residual.mean}
\alias{plot.restricted.residual.mean}
\alias{summary.resmean}
\alias{plot.resmean}
\alias{print.resmean}
\alias{coef.resmean}
\alias{kmplot}
\alias{pval}
\alias{additive.compSs}
\alias{CsmoothB}
\alias{mypbc}
\alias{Csmooth2B}
\alias{pval}
\alias{Event}
\alias{as.character.Event}
\alias{as.matrix.Event} 
\alias{[.Event} 
\alias{format.Event}
\alias{print.Event} 
\alias{rbind.Event} 
\alias{summary.Event}
\alias{kernel}
\alias{percen}
\alias{localTimeReg}
\alias{nameestimate}
\alias{namematrix}
\alias{aalenBase}
\alias{aalenBaseC}
\alias{aalen.des}
\alias{aalen.des2}
\alias{cox.aalenBase}
\alias{read-design}
\alias{is.diag}
\alias{semiaalen}
\alias{semiregBase}
\alias{check.missing} 
\alias{sindex.prodlim} 
\alias{read.design}
\alias{read.surv}
\alias{risk.index}
\alias{faster.reshape}
\alias{rm.missing}
\alias{plot.two.stage}
\alias{summary.two.stage}
\alias{predict.two.stage}
\alias{predictpropodds}
\alias{plot.cums}
\alias{plot.comprisk}
\alias{plot.predict.comprisk}
\alias{plot.predict.timereg}
\alias{pava}
\alias{plotScore}
\alias{summary.comprisk}
\alias{pred.cum}
\alias{slaaop}
\alias{pred.des}
\alias{Cpred}
\alias{plot.cox.aalen2}
\alias{plot.predict}
\alias{coef.aalen}
\alias{cox.marg}
\alias{summary.cox.marg}
\alias{coef.cox.marg}
\alias{print.cox.marg}
\alias{comprisk.ipw}
\alias{coef.comprisk.ipw}
\alias{print.comprisk.ipw}
\alias{prop.odds.subdist.ipw}
\alias{summary.comprisk.ipw}
\alias{coef.cox.aalen}
\alias{coef.comprisk}
\alias{coef.two.stage}
\alias{cluster.index.timereg}
\alias{coefBase}
\alias{des.aalen}
\alias{timetest}
\alias{print.pe.sasieni} 
\alias{pred.stratKM} 
\alias{prep.glm.comprisk} 
\alias{print.predict.timereg}
\alias{print.two.stage}
\alias{residualsTimereg}
\alias{summary.pe.sasieni}
\alias{summary.predict.timereg}
\alias{coef.dynreg}
\alias{coef.timecox}
\alias{dynregBase}
\alias{prop.excessBase}
\alias{prop.odds.gam}
\alias{semicox}
\alias{timecoxBase}
\alias{timereg.formula}
\alias{names2formula}
\alias{twin.clustertrunc}
\alias{vcov.aalen}
\alias{vcov.cox.aalen}
\alias{vcov.two.stage}
\alias{vcov.comp.risk}
\alias{qcut}
\title{For internal use}
\description{for internal use}
\author{Thomas Scheike}
\keyword{survival}
