% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\name{npc}
\alias{CCbinomial.twostage}
\alias{ace.family.design}
\alias{alpha2kendall}
\alias{alpha2spear}
\alias{ascertained.pairs}
\alias{coarse.clust}
\alias{coefmat}
\alias{concordance.twin.ace}
\alias{concordance.twostage}
\alias{corsim.prostate}
\alias{corsim.prostate.random}
\alias{fast.cluster}
\alias{faster.reshape}
\alias{folds}
\alias{force.same.cens}
\alias{grouptable}
\alias{ilap}
\alias{jumptimes}
\alias{kendall.ClaytonOakes.twin.ace}
\alias{kendall.normal.twin.ace}
\alias{loglikMVN}
\alias{make.pairwise.design}
\alias{make.pairwise.design.competing}
\alias{matplot.mets.twostage}
\alias{nonparcuminc}
\alias{npc}
\alias{object.defined}
\alias{p11.binomial.twostage.RV}
\alias{pbvn}
\alias{piecewise.data}
\alias{piecewise.twostage}
\alias{plotcr}
\alias{pmvn}
\alias{predictPairPlack}
\alias{scoreMVN}
\alias{sim}
\alias{simBinFam}
\alias{simBinFam2}
\alias{simBinPlack}
\alias{simClaytonOakes.family.ace}
\alias{simClaytonOakes.twin.ace}
\alias{simCompete.simple}
\alias{simCompete.twin.ace}
\alias{simCox}
\alias{simSurvFam}
\alias{simbinClaytonOakes.family.ace}
\alias{simbinClaytonOakes.pairs}
\alias{simbinClaytonOakes.twin.ace}
\alias{simnordic}
\alias{simnordic.random}
\alias{surv.boxarea}
\alias{twin.polygen.design}
\alias{twostage.fullse}
\title{For internal use}
\description{
For internal use
}
\author{
Klaus K. Holst
}
\keyword{utilities}

