% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R, R/simple-spline.r
\name{LinSpline}
\alias{LinSpline}
\alias{dspline}
\alias{dspline<-}
\alias{LinSpline}
\title{Simple linear spline}
\usage{
LinSpline(x, knots, num = TRUE, signif = NULL, name = "Spline")

LinSpline(x, knots, num = TRUE, signif = NULL, name = "Spline")
}
\arguments{
\item{x}{variable to make into spline}

\item{knots}{cut points}

\item{num}{to give names x1 x2 and so forth}

\item{signif}{to give names x3.2  x5.4 and so forth}

\item{name}{name of spline expansion name.1 name.2 and so forth}

\item{x}{variable to make into spline}

\item{knots}{cut points}

\item{num}{to give names x1 x2 and so forth}
}
\description{
Simple linear spline

Simple linear spline
}
\author{
Thomas Scheike

Thomas Scheike
}
\keyword{survival}
