% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{logitSurv}
\alias{logitSurv}
\title{Fast Proportional odds model}
\usage{
logitSurv(formula, data, offset = NULL, weights = NULL, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph})}

\item{data}{data frame}

\item{offset}{offsets for cox model}

\item{weights}{weights for Cox score equations}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Fast Proportional odds model, that has the advantage that 
\deqn{
logit(S(t|x)) = \Lambda(t) + x \beta
}
}
\examples{
data(TRACE)
dcut(TRACE) <- ~.
out1 <- logitSurv(Surv(time,status==9)~vf+chf+strata(wmicat.4),data=TRACE)
summary(out1)

bplot(out1)
}
\author{
Klaus K. Holst, Thomas Scheike
}
