\name{logLik.biprobit}
\alias{logLik.biprobit}
\title{Calculate log-Likelihood of biprobit object}
\usage{
  logLik.biprobit(object, indiv = FALSE, ...)
}
\arguments{
  \item{object}{}

  \item{indiv}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

