\name{logLik.biprobit.strata}
\alias{logLik.biprobit.strata}
\title{Extract log-Likelihood of biprobit.strata objects}
\usage{
  logLik.biprobit.strata(object, indiv = FALSE,
    list = FALSE, ...)
}
\arguments{
  \item{object}{}

  \item{indiv}{}

  \item{list}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

