\docType{package}
\name{mets-package}
\alias{mets-package}
\title{Analysis of Multivariate Events}
\description{
  Implementation of various statistical models for
  multivariate event history data. Including multivariate
  cumulative incidence models, and bivariate random effects
  probit models (Liability models)
}
\examples{
## To appear
}
\author{
  Klaus K. Holst and Thomas Scheike
}
\keyword{package}

