\name{nonparcuminc}
\alias{nonparcuminc}
\title{IPCW based estimator of the cumulative incidence function}
\usage{
  nonparcuminc(t, status, cens = 0)
}
\arguments{
  \item{t}{}

  \item{status}{}

  \item{cens}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

