\name{npc}
\alias{npc}
\title{IPCW based estimator of the bivariate cumulative incidence function}
\usage{
  npc(T, cause, same.cens = TRUE, sep = FALSE)
}
\arguments{
  \item{T}{}

  \item{cause}{}

  \item{same.cens}{}

  \item{sep}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

