% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phreg}
\alias{phreg}
\alias{phreg.par}
\title{Fast Cox PH regression}
\usage{
phreg(formula, data, offset = NULL, weights = NULL, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph})}

\item{data}{data frame}

\item{weights}{weights for Cox score equations}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Fast Cox PH regression
}
\examples{
data(TRACE)
dcut(TRACE) <- ~.
out1 <- phreg(Surv(time,status==9)~vf+chf+strata(wmicat.4),data=TRACE)
tracesim <- sim.cox(out1,1000)
sout1 <- phreg(Surv(time,status==1)~vf+chf+strata(wmicat.4),data=tracesim)

par(mfrow=c(1,2))
baseplot.phreg(out1)
baseplot.phreg(sout1)
}
\author{
Klaus K. Holst, Thomas Scheike
}

