\name{plack.cif}
\alias{plack.cif}
\alias{plack.cif2}
\title{plack Computes concordance for or.cif based model, that is Plackett random effects model}
\usage{
  plack.cif(cif1, cif2, object)

  plack.cif(cif1, cif2, object)
}
\arguments{
  \item{cif1}{Cumulative incidence of first argument.}

  \item{cif2}{Cumulative incidence of second argument.}

  \item{object}{or.cif object with dependence parameters.}

  \item{cif1}{Cumulative incidence of first argument.}

  \item{cif2}{Cumulative incidence of second argument.}

  \item{object}{or.cif object with dependence parameters.}
}
\description{
  .. content for description (no empty lines) ..

  .. content for description (no empty lines) ..
}
\author{
  Thomas Scheike

  Thomas Scheike
}

