\name{plot.biprobit.strata}
\alias{plot.biprobit.strata}
\title{Plot function for biprobit.strata objects}
\usage{
  plot.biprobit.strata(x, ...)
}
\arguments{
  \item{x}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

