\name{plot.bptwin}
\alias{plot.bptwin}
\title{Plot function for bptwin}
\usage{
  \method{plot}{bptwin} (x, n = 50, rg = range(x$B[, 1]),
    xlab = "Time", ylab = "Concordance", ...)
}
\arguments{
  \item{x}{}

  \item{n}{}

  \item{rg}{}

  \item{xlab}{}

  \item{ylab}{}

  \item{...}{}
}
\description{
  ##' .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

