\name{plotcr}
\alias{plotcr}
\title{Plot bivariate competing risks data}
\usage{
  plotcr(x, col, lty, legend = TRUE, which = 1:2,
    ask = prod(par("mfcol")) < length(which) && dev.interactive(),
    ...)
}
\arguments{
  \item{x}{}

  \item{col}{}

  \item{lty}{}

  \item{legend}{}

  \item{which}{}

  \item{ask}{}

  \item{...}{}
}
\description{
  .. content for description (no empty lines) ..
}
\details{
  .. content for details ..
}
\author{
  Klaus K. Holst
}

