% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pmvn}
\alias{loglikMVN}
\alias{pbvn}
\alias{pmvn}
\alias{scoreMVN}
\title{Multivariate normal distribution function}
\usage{
pmvn(lower, upper, mu, sigma, cor = FALSE)
}
\description{
Multivariate normal distribution function
}
\examples{
lower <- rbind(c(0,-Inf),c(-Inf,0))
upper <- rbind(c(Inf,0),c(0,Inf))
mu <- rbind(c(1,1),c(-1,1))
sigma <- diag(2)+1
pmvn(lower=lower,upper=upper,mu=mu,sigma=sigma)
}
\seealso{
dmvn rmvn
}

