% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.phreg}
\alias{covfr}
\alias{covfridstrata}
\alias{covfridstrataCov}
\alias{cumsumidstratasum}
\alias{cumsumidstratasumCov}
\alias{cumsumstrata}
\alias{cumsumstratasum}
\alias{predict.phreg}
\alias{revcumsumidstratasum}
\alias{revcumsumidstratasumCov}
\alias{revcumsumstrata}
\alias{revcumsumstratasum}
\alias{robust.basehaz.phreg}
\alias{sumstrata}
\title{Predictions from proportional hazards model}
\usage{
\method{predict}{phreg}(object, data, surv = FALSE, time = object$exit,
  X = object$X, strata = object$strata, ...)
}
\arguments{
\item{object}{phreg object}

\item{data}{data.frame}

\item{surv}{If TRUE predictions are provided on probability scale}

\item{time}{Time variable}

\item{X}{Design matrix}

\item{strata}{Strata variable}

\item{...}{ADditional arguments to lower level functions}
}
\description{
Predictions from proportional hazards model
}

