\name{print.aalen}
\alias{print.aalen}
\alias{print.cox.aalen}
\alias{print.comprisk}
\alias{print.prop.excess}
\alias{print.dynreg}
\alias{print.timecox}
\alias{print.cum.residuals}
\title{Prints call}
\description{Prints call for object. Lists nonparametric and 
parametric terms of model}
\usage{\method{print}{aalen}(x,...)}
\arguments{
  \item{x}{an aalen object}
  \item{...}{unused arguments - for S3 compatibility}
}
\author{Thomas Scheike}
\keyword{survival}
