\name{print.biprobit}
\alias{print.biprobit}
\title{Print function for biprobit objects}
\usage{
  print.biprobit(x, ...)
}
\arguments{
  \item{x}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

