% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.casewise}
\alias{print.casewise}
\title{prints Concordance test}
\usage{
\method{print}{casewise}(x, digits = 3, ...)
}
\arguments{
\item{x}{output from casewise.test}

\item{digits}{number of digits}

\item{\dots}{Additional arguments to lower level functions}
}
\description{
prints Concordance test
}
\author{
Thomas Scheike
}

