\name{print.casewisetest}
\alias{print.casewisetest}
\title{prints Concordance test}
\usage{
  print.casewisetest(x, digits = 3, ...)

  print.casewisetest(x, digits = 3, ...)
}
\arguments{
  \item{x}{output from casewise.test}

  \item{digits}{number of digits}

  \item{x}{output from casewise.test}

  \item{digits}{number of digits}
}
\description{
  .. content for description (no empty lines) ..

  .. content for description (no empty lines) ..
}
\author{
  Thomas Scheike

  Thomas Scheike
}

