% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prob.exceed.recurrent}
\alias{prob.exceed.recurrent}
\title{Estimation of probability of more that k events for recurrent events process}
\usage{
prob.exceed.recurrent(data, type, status = "status", death = "death",
  start = "start", stop = "stop", id = "id", times = NULL,
  exceed = NULL)
}
\arguments{
\item{data}{data-frame}

\item{type}{type of evnent (code) related to status}

\item{status}{name of status}

\item{death}{name of death indicator}

\item{start}{start stop call of Hist() of prodlim}

\item{stop}{start stop call of Hist() of prodlim}

\item{id}{}

\item{times}{time at which to get probabilites P(N1(t) >= n)}

\item{exceed}{n's for which which to compute probabilites P(N1(t) >= n)}

\item{max.recurrent}{limits number recurrent events to 100}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Estimation of probability of more that k events for recurrent events process
where there is terminal event
}
\examples{
\donttest{
### do not test to avoid dependence on prodlim
library(prodlim)

pp <- prob.exceed.recurrent(r,1,status="event",death="death",start="t.start",stop="t.stop",id="id")
with(pp, matplot(times,prob,type="s"))
###
with(pp, matlines(times,se.lower,type="s"))
with(pp, matlines(times,se.upper,type="s"))
}
}
\author{
Thomas Scheike
}

