\name{prop}
\alias{prop}
\title{Identifies the multiplicative terms in Cox-Aalen model and
proportional excess risk model}
\description{Specifies which  of the regressors that belong to the
multiplicative part of the Cox-Aalen model 

\deqn{ \lambda_{i}(t) = Y_i(t)
( X_{i}^T(t) \alpha(t) )  \exp(Z_{i}^T(t) \beta )
}
for this model prop specified the covariates to be included in \eqn{Z_{i}(t)}
}
\author{Thomas Scheike}
\keyword{survival}
