\name{score.biprobit}
\alias{score.biprobit}
\title{Extract score of biprobit object}
\usage{
  score.biprobit(x, indiv = FALSE, ...)
}
\arguments{
  \item{x}{}

  \item{indiv}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

