\name{score.biprobit.strata}
\alias{score.biprobit.strata}
\title{Extract score of biprobit.strata objects}
\usage{
  score.biprobit.strata(x, ...)
}
\arguments{
  \item{x}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

