\name{sim.clayton.oakes}
\alias{sim.clayton.oakes}
\title{Simulate from the Clayton-Oakes frailty model}
\usage{
  sim.clayton.oakes(K, n, eta, beta, stoptime, left = 0)
}
\arguments{
  \item{K}{}

  \item{n}{}

  \item{eta}{}

  \item{beta}{}

  \item{stoptime}{}

  \item{left}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

