% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{sim.recurrentII}
\alias{sim.recurrentII}
\title{Simulation of recurrent events data based on cumulative hazards}
\usage{
\method{sim}{recurrentII}(n, cumhaz, cumhaz2, death.cumhaz = NULL,
  gap.time = FALSE, max.recurrent = 100, dhaz = NULL, haz2 = NULL,
  dependence = 0, var.z = 0.4, cor.mat = NULL, ...)
}
\arguments{
\item{cumhaz}{cumulative hazard of recurrent events}

\item{cumhaz2}{cumulative hazard of recurrent events  of type 2}

\item{gap.time}{if true simulates gap-times with specified cumulative hazard}

\item{max.recurrent}{limits number recurrent events to 100}

\item{dhaz}{rate for death hazard if it is extended to time-range of first event}

\item{haz2}{rate of second cause  if it is extended to time-range of first event}

\item{dependence}{=0 independence, =1 all share same random effect with variance var.z
=2 random effect exp(normal) with correlation structure from cor.mat,
first random effect is z1 and for N1
second random effect is z2 and for N2
third random effect is for death}

\item{var.z}{variance of random effects}

\item{cor.mat}{correlation matrix for var.z variance of random effects}

\item{...}{Additional arguments to lower level funtions}

\item{cumhaz.death}{cumulative hazard of death}
}
\description{
Simulation of recurrent events data based on cumulative hazards
}
\details{
Must give hazard of death and two recurrent events.  Possible with two
event types and their dependence can be specified but the two recurrent events need
to share random effect. Based on drawing the from cumhaz and cumhaz2 and taking the first event rather
the cumulative and then distributing it out. Key advantage of this is that there is 
more flexibility wrt random effects
}
\examples{

########################################
## getting some data to mimick 
########################################

 data(simrecurrent)
 xr <- phreg(Surv(start,stop,status)~cluster(id),data=simd)
 dr <- phreg(Surv(start,stop,death)~cluster(id),data=simd)
 par(mfrow=c(1,3))
 basehazplot.phreg(dr,se=TRUE)
 title(main="death")
 basehazplot.phreg(xr,se=TRUE)
 ### robust standard errors 
 rxr <-   robust.phreg(xr,fixbeta=1)
 basehazplot.phreg(rxr,se=TRUE,robust=TRUE,add=TRUE,col=4)
 meanr <-   recurrent.marginal(xr,dr)
 basehazplot.phreg(meanr,se=TRUE)

 var.gamma <- c(0.2,0.5,0.5)
 cor.mat <- corM <- rbind(c(1.0, 0.6, 0.9), c(0.6, 1.0, 0.5), c(0.9, 0.5, 1.0))

######################################################################
### simulating simple model that mimicks data 
### now with two event types and second type has same rate as death rate
######################################################################

 rr <- sim.recurrentII(1000,xr$cumhaz,dr$cumhaz,death.cumhaz=dr$cumhaz)
 dtable(rr,~death+status)
 par(mfrow=c(2,2))
 drr <- phreg(Surv(entry,time,death)~cluster(id),data=rr)
 basehazplot.phreg(drr,se=TRUE)
 basehazplot.phreg(dr,add=TRUE,col=2)
 ###
 xrr1 <- phreg(Surv(entry,time,status==1)~cluster(id),data=rr)
 xrr2 <- phreg(Surv(entry,time,status==2)~cluster(id),data=rr)
 basehazplot.phreg(xrr1,se=TRUE)
 basehazplot.phreg(xr,add=TRUE,col=2)
 basehazplot.phreg(xrr2)
 basehazplot.phreg(dr,col=2,add=TRUE)
 meanr1 <-   recurrent.marginal(xrr1,drr)
 meanr2 <-   recurrent.marginal(xrr2,drr)
 basehazplot.phreg(meanr1,se=TRUE)
 basehazplot.phreg(meanr,se=TRUE,add=TRUE,col=2)
 basehazplot.phreg(meanr2,se=TRUE,add=TRUE)


}
\author{
Thomas Scheike
}
