% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.clayton.oakes.R
\name{simClaytonOakes}
\alias{simClaytonOakes}
\alias{simClaytonOakesLam}
\title{Simulate from the Clayton-Oakes frailty model}
\usage{
simClaytonOakes(K, n, eta, beta, stoptime, lam = 1, left = 0,
  pairleft = 0, trunc.prob = 0.5, same = 0)
}
\arguments{
\item{K}{Number of clusters}

\item{n}{Number of observations in each cluster}

\item{eta}{variance}

\item{beta}{Effect (log hazard ratio) of covariate}

\item{stoptime}{Stopping time}

\item{lam}{constant hazard}

\item{left}{Left truncation}

\item{pairleft}{pairwise (1) left truncation or individual (0)}

\item{trunc.prob}{Truncation probability}

\item{same}{if 1 then left-truncation is same also for univariate truncation}
}
\description{
Simulate observations from the Clayton-Oakes copula model with
piecewise constant marginals.
}
\author{
Thomas Scheike and Klaus K. Holst
}
