\name{summary.biprobit}
\alias{summary.biprobit}
\title{Summary function for biprobit objects}
\usage{
  summary.biprobit(object, level = 0.05, ...)
}
\arguments{
  \item{object}{}

  \item{level}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

