\name{summary.biprobit.strata}
\alias{summary.biprobit.strata}
\title{Summary function for biprobit.strata objects}
\usage{
  summary.biprobit.strata(object, ...)
}
\arguments{
  \item{object}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

