\name{summary.bptwin}
\alias{print.summary.bptwin}
\alias{summary.bptwin}
\title{Summary of bptwin objects}
\usage{
  summary.bptwin(object, level = 0.05, ...)
}
\arguments{
  \item{object}{}

  \item{level}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

