\name{summary.cum.residuals}
\alias{summary.cum.residuals}
\title{Prints summary statistics for goodness-of-fit tests based on cumulative residuals}
\description{
Computes p-values for extreme behaviour relative to the model of 
various cumulative residual processes.}
\usage{\method{summary}{cum.residuals}(object,digits=3,...) }
\arguments{
\item{object}{output from the cum.residuals() function.}
\item{digits}{number of digits in printouts.}
\item{...}{unused arguments - for S3 compatibility}
}
\author{Thomas Scheike}
\examples{
# see cum.residuals for examples
}
\keyword{survival}
