% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifetable.R
\name{survpois}
\alias{survpois}
\title{Extract survival estimates from lifetable analysis}
\usage{
survpois(object, ..., timevar = "int.end", time, int.len, confint = FALSE,
  level = 0.95, individual = FALSE, length.out = 25)
}
\arguments{
\item{object}{glm object (poisson regression)}

\item{...}{Contrast arguments}

\item{timevar}{Name of time variable}

\item{time}{Time points (optional)}

\item{int.len}{Time interval length (optional)}

\item{confint}{If TRUE confidence limits are supplied}

\item{level}{Level of confidence limits}

\item{individual}{Individual predictions}

\item{length.out}{Length of time vector}
}
\description{
Summary for survival analyses via the 'lifetable' function
}
\details{
Summary for survival analyses via the 'lifetable' function
}
\author{
Klaus K. Holst
}

