% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{twinstut}
\alias{twinstut}
\title{Stutter data set}
\format{tvparnr: twin-pair id
zyg: zygosity, MZ:=mz, DZ(same sex):=dz, DZ(opposite sex):=os
stutter: stutter status (yes/no)
age: age
nr: number within twin-pair}
\description{
Based on nation-wide questionnaire answers from 33,317 Danish twins
}
\keyword{data}

