\name{vcov.biprobit}
\alias{vcov.biprobit}
\title{Extract variance covariance function of biprobit object}
\usage{
  vcov.biprobit(object, ...)
}
\arguments{
  \item{object}{}

  \item{...}{}
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Klaus Kähler Holst
}

