\name{back2timereg}
\alias{back2timereg}
\title{Convert to timereg object}
\usage{
  back2timereg(obj)

  back2timereg(obj)
}
\arguments{
  \item{obj}{no use}

  \item{obj}{no use}
}
\description{
  convert to timereg object

  convert to timereg object
}
\author{
  Thomas Scheike

  Thomas Scheike
}

