% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{base2cumhaz}
\alias{base2cumhaz}
\title{rate of Mechanical complication for catheter of HPN patients of Copenhagen}
\source{
Estimated data
}
\description{
rate of Mechanical complication for catheter of HPN patients of Copenhagen
}
\keyword{data}

