\name{bicomprisk}
\alias{bicomprisk}
\title{Bivariate Concordance Estimator}
\usage{
  bicomprisk(formula, data, cause = c(1, 1), cens = 0,
    causes, indiv, strata = NULL, id, num, prodlim = FALSE,
    messages = TRUE, model, return.data = 0, uniform = 0,
    robust = 0, ...)
}
\arguments{
  \item{formula}{}

  \item{data}{}

  \item{cause}{}

  \item{cens}{}

  \item{causes}{}

  \item{indiv}{}

  \item{strata}{}

  \item{id}{}

  \item{num}{}

  \item{prodlim}{}

  \item{messages}{}

  \item{model}{}

  \item{return.data}{}

  \item{uniform}{}

  \item{robust}{}

  \item{...}{}
}
\description{
  .. content for description (no empty lines) ..
}
\details{
  .. content for details ..
}
\author{
  Thomas Scheike, Klaus K. Holst
}

