\name{conc2case}
\alias{conc2case}
\title{Concordance to casewise}
\usage{
  conc2case(conc, marg, test = "conc")
}
\arguments{
  \item{conc}{Concordance}

  \item{marg}{Marginal estimate}

  \item{test}{Type of test for independence assumption.
  "conc" makes test on concordance scale and "case" means a
  test on the probandwise concordance}
}
\description{
  .. content for description (no empty lines) ..
}
\author{
  Thomas Scheike
}

