% Generated by roxygen2 (4.0.1): do not edit by hand
\name{concordance}
\alias{concordance}
\title{Concordance Computes concordance and probandwise concordance}
\usage{
concordance(object, cif1, cif2 = NULL, messages = TRUE, model = NULL,
  coefs = NULL, ...)
}
\arguments{
\item{object}{Output from the cor.cif, rr.cif or or.cif function}

\item{cif1}{Marginal cumulative incidence}

\item{cif2}{Marginal cumulative incidence of other cause (cause2) if  it is different from cause1}

\item{messages}{To print messages}

\item{model}{Specfifies wich model that is considered if object not given.}

\item{coefs}{Specfifies  dependence parameters if object is not given.}

\item{...}{Extra arguments, not used.}
}
\description{
Concordance
}
\author{
Thomas Scheike
}

