% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{daggregate}
\alias{daggregate}
\title{aggregating for for data frames}
\usage{
daggregate(data, y = NULL, x = NULL, ..., fun = "summary",
  regex = FALSE, silent = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{y}{name of variable, or formula, or names of variables on data frame.}

\item{x}{name of variable, or formula, or names of variables on data frame.}

\item{...}{additional arguments to lower level functions}

\item{fun}{function defining aggregation}

\item{regex}{interpret x,y as regular expressions}

\item{silent}{suppress messages}
}
\description{
aggregating for for data frames
}
\examples{
data("sTRACE",package="timereg")
daggregate(iris, "^.e.al", x="Species", fun=cor, regex=TRUE)
daggregate(iris, Sepal.Length+Petal.Length ~Species, fun=summary)
daggregate(iris, log(Sepal.Length)+I(Petal.Length>1.5) ~ Species, fun=summary)
daggregate(iris, "*Length*", x="Species", fun=head)
daggregate(iris, "^.e.al", x="Species", fun=tail, regex=TRUE)
daggregate(sTRACE, status~ diabetes, fun=table)
daggregate(sTRACE, status~ diabetes+sex, fun=table)
daggregate(sTRACE, status + diabetes+sex ~ vf+I(wmi>1.4), fun=table)
daggregate(iris, "^.e.al", x="Species",regex=TRUE)
dprint(iris,Petal.Length+Sepal.Length ~ Species |Petal.Length>1.3 & Sepal.Length>5, n=list(1:3,-(3:1)))
daggregate(iris, I(Sepal.Length>7)~Species | I(Petal.Length>1.5))
daggregate(iris, I(Sepal.Length>7)~Species | I(Petal.Length>1.5), fun=table)

}

