% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dby.R
\name{dby}
\alias{dby}
\title{Calculate summary statistics grouped by}
\usage{
dby(data, input.variable, ..., id.variable = NULL, order.variable = NULL,
  sort.order = 0, combine.data = TRUE, no.check = FALSE, args = NULL,
  names)
}
\arguments{
\item{data}{Data.frame}

\item{input.variable}{Input variables (character or formula)}

\item{...}{functions}

\item{id.variable}{id}

\item{order.variable}{order}

\item{sort.order}{sort order (id+order variable)}

\item{combine.data}{If TRUE result is appended to data}

\item{no.check}{No sorting or check for missing data}

\item{args}{Optional list of arguments to functions (...)}

\item{names}{Optional vector of column names}
}
\description{
Calculate summary statistics grouped by variable
}
\details{
Calculate summary statistics grouped by
}
\examples{
n <- 4
k <- c(3,rbinom(n-1,3,0.5)+1)
N <- sum(k)
d <- data.frame(y=rnorm(N),x=rnorm(N),id=rep(seq(n),k),num=unlist(sapply(k,seq)))
d2 <- d[sample(nrow(d)),]

dby(d, y~id, mean)
dby(d, y~id|num, cumsum)

dby(d,y~id|num, dlag, args=list(k=1:2))

dby(d, y~id|num, dlag)
dby(d, y~id|num, y1=dlag, args=list(k=1:2), names=c("y1","y2"))
dby(d, y~id|num, mean=mean, csum=cumsum, n=length)
dby(d2,y~id|num, a=cumsum, b=mean, N=length, l1=function(x) c(NA,x)[-length(x)])

dby(d, y~id|num, nn=seq, n=length)

f <- function(x) apply(x,1,min)
dby(d, y+x~id, min=f)

dby(d,y+x~id|num, function(x) x)

f <- function(x) { cbind(cumsum(x[,1]),cumsum(x[,2]))/sum(x)}
dby(d, y+x~id, f)
}
\author{
Klaus K. Holst and Thomas Scheike
}

