% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daggregate.R
\name{dcor}
\alias{dcor}
\alias{dcount}
\alias{deval}
\alias{deval2}
\alias{dmean}
\alias{dmeansd}
\alias{dquantile}
\alias{dscalar}
\alias{dsd}
\alias{dstr}
\alias{dsubset}
\alias{dsum}
\alias{dsummary}
\title{summary, tables, and correlations for data frames}
\usage{
dcor(data, y = NULL, x = NULL, ...)
}
\arguments{
\item{data}{if x is formula or names for data frame then data frame is needed.}

\item{y}{name of variable, or fomula, or names of variables on data frame.}

\item{x}{possible group variable}

\item{...}{Optional additional arguments}
}
\description{
summary, tables, and correlations for data frames
}
\examples{
data("sTRACE",package="timereg")
dt<- sTRACE
dt$time2 <- dt$time^2
dt$wmi2 <- dt$wmi^2
head(dt)

dcor(dt)

dcor(dt,~time+wmi)
dcor(dt,~time+wmi,~vf+chf)
dcor(dt,time+wmi~vf+chf)

dcor(dt,c("time*","wmi*"),~vf+chf)
}
\author{
Klaus K. Holst and Thomas Scheike
}

