% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{dermalridgesMZ}
\alias{dermalridgesMZ}
\title{Dermal ridges data (monozygotic twins)}
\format{Data on dermal ridge counts (left and right hand) in 18
monozygotic twin pairs.}
\source{
Sarah B. Holt (1952). Genetics of dermal ridges: bilateral
asymmetry in finger ridge-counts.  Annals of Eugenics 17 (1),
pp.211--231. DOI: 10.1111/j.1469-1809.1952.tb02513.x
}
\description{
Data on dermal ridge counts in left and right hand in (nuclear) families
}
\examples{
data(dermalridgesMZ)
fast.reshape(dermalridgesMZ,id="id",varying=c("left","right"))
}
\keyword{data}

