% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/divide.conquer.R
\name{divide.conquer}
\alias{divide.conquer}
\title{Split a data set and run function}
\usage{
divide.conquer(func = NULL, data, size, ...)
}
\arguments{
\item{func}{called function}

\item{data}{data-frame}

\item{size}{size of splits}

\item{...}{Additional arguments to lower level functions}
}
\description{
Split a data set and run function
}
\examples{
library(timereg)
data(TRACE)
\donttest{
res <- divide.conquer(prop.odds,TRACE,
	     formula=Surv(time,status==9)~chf+vf+age,n.sim=0,size=200)
}
}
\author{
Thomas Scheike, Klaus K. Holst
}

