% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{familycluster.index}
\alias{familycluster.index}
\title{Finds all pairs within a cluster (family)}
\usage{
familycluster.index(clusters, index.type = FALSE, num = NULL, Rindex = 1)
}
\description{
Finds all pairs within a cluster (family)
}
\examples{
i<-c(1,1,2,2,1,3)
d<- familycluster.index(i)
print(d)
}
\author{
Klaus Holst, Thomas Scheike
}
\references{
Cluster indeces
}
\keyword{cluster}
\keyword{indeces}

