% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{familyclusterWithProbands.index}
\alias{familyclusterWithProbands.index}
\title{Finds all pairs within a cluster (famly)  with the proband (case/control)}
\usage{
familyclusterWithProbands.index(clusters, probands, index.type = FALSE,
  num = NULL, Rindex = 1)
}
\arguments{
\item{clusters}{list of indeces giving the clusters (families)}

\item{probands}{list of 0,1 where 1 specifices which of the subjects that are probands}

\item{index.type}{argument passed to other functions}

\item{num}{argument passed to other functions}

\item{Rindex}{index starts with 1, in C is it is 0}
}
\description{
second column of pairs are the probands and the first column the related subjects
}
\examples{
i<-c(1,1,2,2,1,3)
p<-c(1,0,0,1,0,1)
d<- familyclusterWithProbands.index(i,p)
print(d)
}
\references{
Cluster indeces
}
\seealso{
familycluster.index cluster.index
}
\author{
Klaus Holst, Thomas Scheike
}
