\name{fast.approx}
\alias{fast.approx}
\title{Fast approximation}
\usage{
  fast.approx(time, new.time, equal = FALSE, ...)
}
\arguments{
  \item{time}{Original ordered time points}

  \item{new.time}{New time points}

  \item{equal}{If TRUE a list is returned with additional
  element}

  \item{...}{Optional additional arguments}
}
\description{
  Fast approximation
}
\author{
  Klaus K. Holst
}

