% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fast.approx}
\alias{fast.approx}
\title{Fast approximation}
\usage{
fast.approx(time, new.time, equal = FALSE, ...)
}
\arguments{
\item{time}{Original ordered time points}

\item{new.time}{New time points}

\item{equal}{If TRUE a list is returned with additional element}

\item{...}{Optional additional arguments}
}
\description{
Fast approximation
}
\examples{
id <- c(1,1,2,2,7,7,10,10)
fast.approx(unique(id),id)

t <- 0:6
n <- c(-1,0,0.1,0.9,1,1.1,1.2,6,6.5)
fast.approx(t,n,equal=TRUE)
}
\author{
Klaus K. Holst
}

