% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fast.pattern}
\alias{fast.pattern}
\title{Fast pattern}
\usage{
fast.pattern(x, y, categories = 2, ...)
}
\arguments{
\item{x}{Matrix (binary) of patterns. Optionally if \code{y} is
also passed as argument, then the pattern matrix is defined as the
elements agreeing in the two matrices.}

\item{y}{Optional matrix argument with same dimensions as \code{x} (see above)}

\item{categories}{Default 2 (binary)}

\item{...}{Optional additional arguments}
}
\description{
Fast pattern
}
\examples{
X <- matrix(rbinom(100,1,0.5),ncol=4)
fast.pattern(X)

X <- matrix(rbinom(100,3,0.5),ncol=4)
fast.pattern(X,categories=4)
}
\author{
Klaus K. Holst
}

