\name{pval}
\alias{pval}
\alias{cv.folds}
\alias{additive.compSs}
\alias{CsmoothB}
\alias{mypbc}
\alias{Csmooth2B}
\alias{pval}
\alias{kernel}
\alias{percen}
\alias{localTimeReg}
\alias{nameestimate}
\alias{namematrix}
\alias{aalen.des}
\alias{aalenBase}
\alias{aalenBaseC}
\alias{cox.aalenBase}
\alias{prop.excessBase}
\alias{read-design}
\alias{dynregBase}
\alias{timecoxBase}
\alias{semiaalen}
\alias{semicox}
\alias{semiregBase}
\alias{check.missing} 
\alias{read.design}
\alias{read.surv}
\alias{rm.missing}
\alias{plot.two.stage}
\alias{summary.two.stage}
\alias{summary.two.stage.reg}
\alias{two.stageBase}
\alias{two.stageBase.reg}
\alias{plot.cums}
\alias{pava}
\alias{plotScore}
\alias{summary.comprisk}
\alias{pred.cum}
\alias{slaaop}
\alias{pred.des}
\alias{Cpred}
\alias{plot.cox.aalen2}
\alias{plot.prop.excess2}
\alias{plot.timecox2}
\alias{plot.predict}
\alias{coef.aalen}
\alias{coef.cox.aalen}
\alias{coef.comprisk}
\alias{coef.dynreg}
\alias{coef.timecox}
\alias{coef.two.stage}
\alias{coefBase}
\alias{des.aalen}
\alias{timetest}
\alias{plot.comprisk} 
\alias{plot.predictAalen} 
\alias{plot.predictComprisk}
\alias{plot.predictCoxAalen} 
\alias{print.pe.sasieni} 
\alias{print.predictAalen}
\alias{print.predictComprisk} 
\alias{print.predictCoxAalen}
\alias{print.two.stage}
\alias{print.two.stage.reg} 
\alias{summary.pe.sasieni}
\alias{summary.predictAalen}
\alias{summary.predictComprisk} 
\alias{summary.predictCoxAalen}
\alias{summary.pls}
\alias{print.pls}
\alias{predict.pls}
\alias{my.lars}
\alias{myplot.lars}
\title{For internal use}
\description{for internal use}
\author{Thomas Scheike}
\keyword{survival}
