\name{pval}
\alias{pval}
\alias{additive.compSs}
\alias{CsmoothB}
\alias{mypbc}
\alias{Csmooth2B}
\alias{pval}
\alias{kernel}
\alias{percen}
\alias{localTimeReg}
\alias{nameestimate}
\alias{namematrix}
\alias{aalenBase}
\alias{aalenBaseC}
\alias{aalen.des}
\alias{aalen.des2}
\alias{cox.aalenBase}
\alias{read-design}
\alias{is.diag}
\alias{semiaalen}
\alias{semiregBase}
\alias{check.missing} 
\alias{read.design}
\alias{read.surv}
\alias{risk.index}
\alias{faster.reshape}
\alias{rm.missing}
\alias{plot.two.stage}
\alias{summary.two.stage}
\alias{predict.two.stage}
\alias{predictpropodds}
\alias{plot.cums}
\alias{plot.comprisk}
\alias{plot.predict.comprisk}
\alias{plot.predict.timereg}
\alias{pava}
\alias{plotScore}
\alias{summary.comprisk}
\alias{pred.cum}
\alias{slaaop}
\alias{pred.des}
\alias{Cpred}
\alias{plot.cox.aalen2}
\alias{plot.predict}
\alias{coef.aalen}
\alias{coef.cox.aalen}
\alias{coef.comprisk}
\alias{coef.two.stage}
\alias{cluster.index}
\alias{coefBase}
\alias{des.aalen}
\alias{timetest}
\alias{print.pe.sasieni} 
\alias{print.predict.timereg}
\alias{print.two.stage}
\alias{residualsTimereg}
\alias{summary.pe.sasieni}
\alias{summary.predict.timereg}
\alias{coef.dynreg}
\alias{coef.timecox}
\alias{dynregBase}
\alias{prop.excessBase}
\alias{semicox}
\alias{timecoxBase}
\alias{wald.test}
\title{For internal use}
\description{for internal use}
\author{Thomas Scheike}
\keyword{survival}
