\name{npc}
\alias{alpha2kendall}
\alias{alpha2spear}
\alias{cluster.index}
\alias{coefmat}
\alias{corsim.prostate}
\alias{familycluster.index}
\alias{faster.reshape}
\alias{nonparcuminc}
\alias{npc}
\alias{piecewise.data}
\alias{piecewise.twostage}
\alias{plotcr}
\alias{simBinFam}
\alias{simBinPlack}
\alias{simnordic}
\alias{surv.boxarea}
\title{For internal use}
\description{
  For internal use
}
\author{
  Klaus K. Holst
}
\keyword{utilities}

