% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mets-package.R
\name{npc}
\alias{alpha2kendall}
\alias{alpha2spear}
\alias{cluster.index}
\alias{coefmat}
\alias{corsim.prostate}
\alias{familycluster.index}
\alias{faster.reshape}
\alias{folds}
\alias{grouptable}
\alias{jumptimes}
\alias{loglikMVN}
\alias{nonparcuminc}
\alias{npc}
\alias{pbvn}
\alias{piecewise.data}
\alias{piecewise.twostage}
\alias{plotcr}
\alias{pmvn}
\alias{sim}
\alias{simBinFam}
\alias{simBinFam2}
\alias{simBinPlack}
\alias{simCox}
\alias{simSurvFam}
\alias{simnordic}
\alias{surv.boxarea}
\title{For internal use}
\description{
For internal use
}
\author{
Klaus K. Holst
}
\keyword{utilities}

