% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{LinSpline}
\alias{LinSpline}
\title{Simple linear spline}
\usage{
LinSpline(x, knots, num = TRUE, name = "Spline")
}
\arguments{
\item{x}{variable to make into spline}

\item{knots}{cut points}

\item{num}{to give names x1 x2 and so forth}

\item{name}{name of spline expansion name.1 name.2 and so forth}
}
\description{
Simple linear spline
}
\author{
Thomas Scheike
}
\keyword{survival}
