% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{package}
\name{mets-package}
\alias{mets-package}
\title{Analysis of Multivariate Events}
\description{
Implementation of various statistical models for multivariate
event history data. Including multivariate cumulative incidence models,
and bivariate random effects probit models (Liability models)
}
\examples{
## To appear
}
\author{
Klaus K. Holst and Thomas Scheike
}
\keyword{package}

