% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{mets.options}
\alias{mets.options}
\title{Set global options for \code{mets}}
\usage{
mets.options(...)
}
\arguments{
\item{...}{Arguments}
}
\value{
\code{list} of parameters
}
\description{
Extract and set global parameters of \code{mets}.
}
\details{
\itemize{
  \item \code{regex}: If TRUE character vectors will be interpreted as regular expressions (\code{dby}, \code{dcut}, ...)
  \item \code{silent}: Set to \code{FALSE} to disable various output messages
}
}
\examples{
\dontrun{
mets.options(regex=TRUE)
}
}
\keyword{models}

