% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{multcif}
\alias{multcif}
\title{Multivariate Cumulative Incidence Function example data set}
\source{
Simulated data
}
\description{
Multivariate Cumulative Incidence Function example data set
}
\keyword{data}
