% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{predict.phreg}
\alias{predict.phreg}
\alias{revcumsumstrata}
\alias{revcumsumstratasum}
\alias{cumsumstrata}
\alias{sumstrata}
\title{Predictions from proportional hazards model}
\usage{
\method{predict}{phreg}(object, data, surv = FALSE, time = object$exit,
  X = object$X, strata = object$strata, ...)
}
\arguments{
\item{object}{phreg object}

\item{data}{data.frame}

\item{surv}{If TRUE predictions are provided on probability scale}

\item{time}{Time variable}

\item{X}{Design matrix}

\item{strata}{Strata variable}

\item{...}{ADditional arguments to lower level functions}
}
\description{
Predictions from proportional hazards model
}
