% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.aalen.r
\name{print.aalen}
\alias{print.aalen}
\alias{print.cox.aalen}
\alias{print.comprisk}
\alias{print.prop.excess}
\alias{print.dynreg}
\alias{print.timecox}
\alias{print.cum.residuals}
\title{Prints call}
\usage{
\method{print}{aalen}(x, ...)
}
\arguments{
\item{x}{an aalen object}

\item{...}{unused arguments - for S3 compatibility}
}
\description{
Prints call for object. Lists nonparametric and parametric terms of model
}
\author{
Thomas Scheike
}
\keyword{survival}
