% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.cox-aalen.r
\name{prop}
\alias{prop}
\title{Identifies the multiplicative terms in Cox-Aalen model and proportional
excess risk model}
\usage{
prop(x)
}
\arguments{
\item{x}{variable}
}
\description{
Specifies which of the regressors that belong to the multiplicative part of
the Cox-Aalen model
}
\details{
\deqn{ \lambda_{i}(t) = Y_i(t) ( X_{i}^T(t) \alpha(t) ) \exp(Z_{i}^T(t)
\beta ) } for this model prop specified the covariates to be included in
\eqn{Z_{i}(t)}
}
\author{
Thomas Scheike
}
\keyword{survival}
