% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.aalen.r
\name{summary.aalen}
\alias{summary.aalen}
\alias{summary.cox.aalen}
\alias{summary.prop.excess}
\alias{summary.timecox}
\alias{summary.dynreg}
\title{Prints summary statistics}
\usage{
\method{summary}{aalen}(object, digits = 3, ...)
}
\arguments{
\item{object}{an aalen object.}

\item{digits}{number of digits in printouts.}

\item{...}{unused arguments - for S3 compatibility}
}
\description{
Computes p-values for test of significance for nonparametric terms of model,
p-values for test of constant effects based on both supremum and integrated
squared difference.
}
\details{
Returns parameter estimates and their standard errors.
}
\examples{

### see help(aalen)

}
\references{
Martinussen and Scheike,
}
\author{
Thomas Scheike
}
\keyword{survival}
