% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgresid.r
\name{summary.cum.residuals}
\alias{summary.cum.residuals}
\title{Prints summary statistics for goodness-of-fit tests based on cumulative
residuals}
\usage{
\method{summary}{cum.residuals}(object, digits = 3, ...)
}
\arguments{
\item{object}{output from the cum.residuals() function.}

\item{digits}{number of digits in printouts.}

\item{...}{unused arguments - for S3 compatibility}
}
\description{
Computes p-values for extreme behaviour relative to the model of various
cumulative residual processes.
}
\examples{

# see cum.residuals for examples

}
\author{
Thomas Scheike
}
\keyword{survival}
